requirements.txt
streamlit
requests
pandas
numpy
pdfplumber
python-pptx
python-docx
sentence-transformers
torch
transformers
faiss-cpu
langchain